+++ FrensHelpers.cpp
@@ -7,6 +7,12 @@
 #include "hardware/flash.h"
 #include "hardware/watchdog.h"
 #include "util/exclusive_proc.h"
+#if CFG_TUH_RPI_PIO_USB
+#include "bsp/board_api.h"
+#include "board.h"
+#include "pio_usb.h"
+#endif
+#include "tusb.h"
 #include "tusb.h"
 #include "dvi/dvi.h"
 #include "ff.h"
@@ -17,7 +23,6 @@
 #include "wiipad.h"
 #include "settings.h"
 #include "FrensHelpers.h"
-
 
 // Pico W devices use a GPIO on the WIFI chip for the LED,
 // so when building for Pico W, CYW43_WL_GPIO_LED_PIN will be defined
@@ -251,9 +256,12 @@ namespace Frens
         };
         bool spi_configured = pico_fatfs_set_config(&config);
         // Try first using SPI
-        if (spi_configured) {
+        if (spi_configured)
+        {
             printf("using SPI...");
-        } else {
+        }
+        else
+        {
             // fall back to PIO SPI
             pico_fatfs_config_spi_pio(SDCARD_PIO, pio_claim_unused_sm(SDCARD_PIO, true));
             printf("using SPI PIO...");
@@ -710,6 +718,32 @@ namespace Frens
         return PICO_OK;
     }
 
+    /// @brief Finds an unused DMA channel.
+    /// This function iterates through the available DMA channels (0-11) and returns the first unused channel.
+    /// If no unused channel is found, it will panic.
+    /// @return the number of the unused DMA channel (0-11).
+    int GetUnUsedDMAChan()
+    {
+        // Get an unused DMA channel
+        int dma_chan = -1;
+        printf("Searching for unused DMA channel...");
+        for (int i = 0; i < 12; i++)
+        {
+            if (!dma_channel_is_claimed(i))
+            {
+                dma_chan = i;
+                printf(" found unused DMA channel %d\n", dma_chan);
+                break;
+            }
+        }
+        if (dma_chan == -1)
+        {
+            panic("No unused DMA channel found");
+        }
+        return dma_chan;
+    }
+
+
     void initVintageControllers(uint32_t CPUFreqKHz)
     {
 #if NES_PIN_CLK != -1
@@ -722,7 +756,31 @@ namespace Frens
         wiipad_begin();
 #endif
     }
+    
+    // Initialize the PIO USB board
+    // replaces board_init() in $PICO_SDK_PATH/lib/tinyusb/src/hw/bsp/rp2040/family.c
+    void pio_usb_board_init(void)
+    {
+#if (CFG_TUH_ENABLED && CFG_TUH_RPI_PIO_USB) || (CFG_TUD_ENABLED && CFG_TUD_RPI_PIO_USB)
+        // power on the PIO USB VBUSEN pin if needed.
+#ifdef PICO_DEFAULT_PIO_USB_VBUSEN_PIN
+        gpio_init(PICO_DEFAULT_PIO_USB_VBUSEN_PIN);
+        gpio_set_dir(PICO_DEFAULT_PIO_USB_VBUSEN_PIN, GPIO_OUT);
+        gpio_put(PICO_DEFAULT_PIO_USB_VBUSEN_PIN, PICO_DEFAULT_PIO_USB_VBUSEN_STATE);
+#endif
 
+        // rp2040 use pico-pio-usb for host tuh_configure() can be used to passed pio configuration to the host stack
+        // Note: tuh_configure() must be called before tuh_init()
+        pio_usb_configuration_t pio_cfg = PIO_USB_DEFAULT_CONFIG;
+        // find an unused DMA channel
+        pio_cfg.tx_ch = GetUnUsedDMAChan();
+        // 
+        pio_cfg.pio_rx_num = PIO_USB_USE_PIO;
+        pio_cfg.pio_tx_num = PIO_USB_USE_PIO;
+        pio_cfg.pin_dp = PICO_DEFAULT_PIO_USB_DP_PIN;
+        tuh_configure(BOARD_TUH_RHPORT, TUH_CFGID_RPI_PIO_USB_CONFIGURATION, &pio_cfg);
+#endif
+    }
     void initDVandAudio(int marginTop, int marginBottom, size_t audioBufferSize)
     {
         //
@@ -758,6 +816,7 @@ namespace Frens
         {
             printf("Error initializing LED: %d\n", rc);
         }
+
         // Calculate the address in flash where roms will be stored
         printf("Flash binary start    : 0x%08x\n", &__flash_binary_start);
         printf("Flash binary end      : 0x%08x\n", &__flash_binary_end);
@@ -767,8 +826,8 @@ namespace Frens
         printf("Size program in flash :   %8d bytes (%d) Kbytes\n", &__flash_binary_end - &__flash_binary_start, (&__flash_binary_end - &__flash_binary_start) / 1024);
         // round ROM_FILE_ADDRESS address up to 4k boundary of flash_binary_end
         ROM_FILE_ADDR = ((uintptr_t)&__flash_binary_end + 0xFFF) & ~0xFFF;
-        //ROM_FILE_ADDR =  0x1004a000;
-        // calculate max rom size
+        // ROM_FILE_ADDR =  0x1004a000;
+        //  calculate max rom size
         maxRomSize = flash_end - (uint8_t *)ROM_FILE_ADDR;
         printf("ROM_FILE_ADDR         : 0x%08x\n", ROM_FILE_ADDR);
         printf("Max ROM size          :   %8d bytes (%d) KBytes\n", maxRomSize, maxRomSize / 1024);
@@ -803,6 +862,26 @@ namespace Frens
             mutex_init(&framebuffer_mutex);
         }
         initDVandAudio(marginTop, marginBottom, audiobufferSize);
+        // init USB driver
+        // USB driver is initalized after display driver to prevent the display driver
+        // from using the PIO state machines already claimed by the USB driver.
+        // This is only needed for the PIO USB driver.
+#if CFG_TUH_RPI_PIO_USB
+        printf("Using PIO USB.\n");
+        pio_usb_board_init();
+        tusb_rhport_init_t host_init = {
+            .role = TUSB_ROLE_HOST,
+            .speed = TUSB_SPEED_AUTO};
+        tusb_init(BOARD_TUH_RHPORT, &host_init);
+
+        if (board_init_after_tusb)
+        {
+            board_init_after_tusb();
+        }
+#else
+        printf("Using internal USB.\n");
+        tusb_init();
+#endif
         if (usingFramebuffer)
         {
             multicore_launch_core1(coreFB_main);
@@ -812,7 +891,7 @@ namespace Frens
             multicore_launch_core1(core1_main);
         }
         initVintageControllers(CPUFreqKHz);
-        EXT_AUDIO_SETUP(DVIAUDIOFREQ);  // Initialize external audio if needed
+        EXT_AUDIO_SETUP(DVIAUDIOFREQ); // Initialize external audio if needed
         return ok;
     }