+++ bld.sh
@@ -13,11 +13,12 @@ APP=${PROJECT}
 function usage() {
 	echo "Build script for the ${PROJECT} project"
 	echo  ""
-	echo "Usage: $0 [-d] [-2 | -r] [-w] [-t path to toolchain] [ -p nprocessors] [-c <hwconfig>]"
+	echo "Usage: $0 [-d] [-2 | -r] [-w] [-u] [-m] [-t path to toolchain] [ -p nprocessors] [-c <hwconfig>]"
 	echo "Options:"
 	echo "  -d: build in DEBUG configuration"
 	echo "  -2: build for Pico 2 board (RP2350)"
 	echo "  -r: build for Pico 2 board (RP2350) with riscv core"
+	echo "  -u: enable PIO USB support (default is disabled)"
 	echo "  -w: build for Pico_w or Pico2_w"
 	echo "  -t <path to riscv toolchain>: only needed for riscv, specify the path to the riscv toolchain bin folder"
 	echo "     Default is \$PICO_SDK_PATH/toolchain/RISCV_RPI_2_0_0_2/bin"
@@ -28,8 +29,9 @@ function usage() {
 	echo "        Custom pcb"
 	echo "     3: Adafruit Feather RP2040 DVI"
 	echo "     4: Waveshare RP2040-PiZero"
-	echo "     5: Adafruit Metro RP2350"
+	echo "     5: Adafruit Metro RP2350 (latest branch of TinyUSB is required for this board)"
 	echo "     6: Waveshare RP2040-Zero/RP2350-Zero with custom PCB"
+	echo "  -m: Run cmake only, do not build the project"
 	echo "  -h: display this help"
 	echo ""
 	echo "To install the RISC-V toolchain:"
@@ -72,7 +74,9 @@ TOOLCHAIN_PATH=
 picoarmIsSet=0
 picoRiscIsSet=0
 USEPICOW=0
-while getopts "whd2rc:t:p:" opt; do
+USEPIOUSB=0
+CMAKEONLY=0
+while getopts "muwhd2rc:t:p:" opt; do
   case $opt in
     p)
 	  BUILDPROC=$OPTARG
@@ -99,6 +103,10 @@ while getopts "whd2rc:t:p:" opt; do
 	  PICO_PLATFORM=rp2350-riscv
 	  ;;	
 	t) TOOLCHAIN_PATH=$OPTARG
+	  ;;
+	u) USEPIOUSB=1
+	  ;;
+	m) CMAKEONLY=1
 	  ;;
 	h)
 	  usage
@@ -176,7 +184,47 @@ if [[ $USEPICOW -eq 1 && $HWCONFIG -gt 2 ]] ; then
 	exit 1
 fi
 
-
+if [[ $HWCONFIG -eq 5 || $USEPIOUSB -eq 1 ]] ; then
+	usbvalid=1
+	if [ ! -d "${PICO_SDK_PATH}/lib/tinyusb/hw/bsp/rp2040/boards/adafruit_metro_rp2350" ] ; then
+		echo "You have not the latest master branch of the TinyUSB library."
+		if [[ $HWCONFIG -eq 5 ]] ; then 
+			 echo "This is needed for the Adafruit Metro RP2350."
+		fi
+		if [[ $USEPIOUSB -eq 1 ]] ; then
+			echo "This is needed for PIO USB support."
+		fi	
+		echo "Please install the latest TinyUSB library:"
+		echo " cd $PICO_SDK_PATH/lib/tinyusb"
+		echo " git checkout master"
+		echo " git pull"
+		usbvalid=0	
+	fi
+	piovalid=1
+	if [[ $USEPIOUSB -eq 1 ]] ; then
+		# check if environment var PICO_PIO_USB_PATH is set and points to a valid path	
+		if [ -z "$PICO_PIO_USB_PATH" ] ; then
+			echo "PICO_PIO_USB_PATH not set."
+			echo "Please set the PICO_PIO_USB_PATH environment variable to the location of the PIO USB library"
+			piovalid=0
+		elif [ ! -r "${PICO_PIO_USB_PATH}/src/pio_usb.h" ] ; then
+			echo "No valid PIO USB repo found."
+			echo "Please set the PICO_PIO_USB_PATH environment variable to the location of the PIO USB library"
+			piovalid=0
+		fi
+		if [ $piovalid -eq 0 ] ; then
+			echo "To install the PIO USB library:"
+			echo " git clone https://github.com/sekigon-gonnoc/Pico-PIO-USB.git"
+			echo " and set the PICO_PIO_USB_PATH environment variable to the location of the Pico-PIO-USB repository"
+			echo " Example: export PICO_PIO_USB_PATH=~/Pico-PIO-USB"
+			echo " or add it to your .bashrc file"
+		fi
+	fi
+	if [[ $piovalid -eq 0 || $usbvalid -eq 0 ]] ; then
+		echo "Please fix the above errors and try again."
+		exit 1
+	fi
+fi
 
 case $HWCONFIG in
 	1)
@@ -210,10 +258,14 @@ fi
 # if [ "$PICO_PLATFORM" = "rp2350-arm-s" ] ; then
 # 	UF2="pico2_$UF2"
 # fi	
+PIOUSB=
+if [ $USEPIOUSB -eq 1 ] ; then
+	PIOUSB="_pio_usb"
+fi	
 if [ "$PICO_PLATFORM" = "rp2350-riscv" ] ; then
-	UF2="${PICO_BOARD}_riscv_$UF2"
+	UF2="${PICO_BOARD}${PIOUSB}_riscv_$UF2"
 else
-	UF2="${PICO_BOARD}_$UF2"
+	UF2="${PICO_BOARD}${PIOUSB}_$UF2"
 fi
 echo "Building $PROJECT"
 echo "Using Pico SDK version: $SDKVERSION"
@@ -228,9 +280,13 @@ fi
 mkdir build || exit 1
 cd build || exit 1
 if [ -z "$TOOLCHAIN_PATH" ] ; then
-	cmake -DCMAKE_BUILD_TYPE=$BUILD -DPICO_BOARD=$PICO_BOARD -DHW_CONFIG=$HWCONFIG -DPICO_PLATFORM=$PICO_PLATFORM .. || exit 1
+	cmake -DCMAKE_BUILD_TYPE=$BUILD -DPICO_BOARD=$PICO_BOARD -DHW_CONFIG=$HWCONFIG -DPICO_PLATFORM=$PICO_PLATFORM -DENABLE_PIO_USB=$USEPIOUSB .. || exit 1
 else
-	cmake -DCMAKE_BUILD_TYPE=$BUILD -DPICO_BOARD=$PICO_BOARD -DHW_CONFIG=$HWCONFIG -DPICO_PLATFORM=$PICO_PLATFORM -DPICO_TOOLCHAIN_PATH=$TOOLCHAIN_PATH .. ||  exit 1
+	cmake -DCMAKE_BUILD_TYPE=$BUILD -DPICO_BOARD=$PICO_BOARD -DHW_CONFIG=$HWCONFIG -DPICO_PLATFORM=$PICO_PLATFORM -DENABLE_PIO_USB=$USEPIOUSB -DPICO_TOOLCHAIN_PATH=$TOOLCHAIN_PATH  .. ||  exit 1
+fi
+if [ $CMAKEONLY -eq 1 ] ; then
+	echo "CMake configuration done, exiting as requested."
+	exit 0
 fi
 make -j $BUILDPROC || exit 1
 cd ..